docs/dev/tutorials/pyproject.toml
[tool.poetry]
name = "tutorials"
version = "0.1.0"
description = ""
authors = ["Bob Steers <steers.bob@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.23.0"
paho-mqtt = "^1.5.0"
pyserial = "^3.4"
schedule = "^0.6.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.2"
pytest = "^5.4.2"
autopep8 = "^1.5.3"
flake8-quotes = "^3.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
